+++ packages/mds-audit-api/tests/api.spec.ts
@@ -16,9 +16,7 @@
 
 /* eslint-disable promise/no-callback-in-promise */
 /* eslint-disable promise/no-nesting */
-/* eslint-disable promise/prefer-await-to-then */
 /* eslint-disable promise/always-return */
-/* eslint-disable promise/prefer-await-to-callbacks */
 /* eslint-disable @typescript-eslint/no-floating-promises */
 /* eslint-disable promise/catch-or-return */
 
@@ -841,7 +839,6 @@ describe('Testing API', () => {
     )
 
     it('verify audit attach (success)', done => {
-      // const fake = Sinon.fake.returns()
       jest.spyOn(attachments, 'writeAttachment').mockImplementation(
         async () =>
           ({