+++ container-images/mds-audit-api/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @container-images/mds-audit-api
 
+## 0.3.0
+
+### Minor Changes
+
+- de647705: Optimize Dockerfiles for caching layers
+
+### Patch Changes
+
+- Updated dependencies [455b9852]
+- Updated dependencies [c9759b66]
+  - @mds-core/mds-audit-api@0.4.4
+  - @mds-core/mds-webpack-config@0.1.13
+
 ## 0.2.1
 
 ### Patch Changes