+++ packages/mds-collector-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mds-core/mds-collector-api
 
+## 0.1.73
+
+### Patch Changes
+
+- Updated dependencies [455b9852]
+- Updated dependencies [c9759b66]
+  - @mds-core/mds-utils@0.5.1
+  - @mds-core/mds-api-server@0.4.12
+  - @mds-core/mds-collector-service@0.4.27
+  - @mds-core/mds-api-helpers@0.2.26
+  - @mds-core/mds-service-helpers@0.6.9
+
 ## 0.1.72
 
 ### Patch Changes