+++ packages/mds-config-service/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mds-core/mds-config-service
 
+## 0.4.10
+
+### Patch Changes
+
+- Updated dependencies [455b9852]
+- Updated dependencies [c9759b66]
+  - @mds-core/mds-utils@0.5.1
+  - @mds-core/mds-rpc-common@0.3.20
+  - @mds-core/mds-config-files@0.1.5
+  - @mds-core/mds-service-helpers@0.6.9
+
 ## 0.4.9
 
 ### Patch Changes