+++ packages/mds-geography-service/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mds-core/mds-geography-service
 
+## 0.10.23
+
+### Patch Changes
+
+- c9759b66: Resolve linter warnings
+- Updated dependencies [455b9852]
+- Updated dependencies [c9759b66]
+  - @mds-core/mds-utils@0.5.1
+  - @mds-core/mds-rpc-common@0.3.20
+  - @mds-core/mds-schema-validators@0.5.11
+  - @mds-core/mds-test-data@0.4.29
+  - @mds-core/mds-repository@0.2.1
+  - @mds-core/mds-service-helpers@0.6.9
+
 ## 0.10.22
 
 ### Patch Changes