+++ packages/mds-policy-service/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @mds-core/mds-policy-service
 
+## 0.12.0
+
+### Minor Changes
+
+- 1e32c275: Rename MDS Policy publish_date -> published_date to conform with MDS 1.0 spec
+
+### Patch Changes
+
+- c9759b66: Resolve linter warnings
+- Updated dependencies [455b9852]
+- Updated dependencies [c9759b66]
+  - @mds-core/mds-stream@0.3.2
+  - @mds-core/mds-utils@0.5.1
+  - @mds-core/mds-geography-service@0.10.23
+  - @mds-core/mds-rpc-common@0.3.20
+  - @mds-core/mds-schema-validators@0.5.11
+  - @mds-core/mds-repository@0.2.1
+  - @mds-core/mds-service-helpers@0.6.9
+
 ## 0.11.28
 
 ### Patch Changes